+++ dataprep-backend-common/src/main/java/org/talend/dataprep/api/dataset/DataSetLocation.java
@@ -42,6 +42,15 @@ public interface DataSetLocation extends Serializable {
     String getLocationType();
 
     /**
+     * @return in case a location has many types (e.g. oracle, mysql, postgres -> database)
+     */
+    @JsonIgnore
+    default String getParentLocationType() {
+        // default is the location type
+        return getLocationType();
+    }
+
+    /**
      * All needed parameters to use this location. Returns {@code null} if not relevant, when {@link #getParametersAsSchema} should
      * be used for instance.
      *
