+++ dataprep-backend-common/src/main/java/org/talend/dataprep/parameters/jsonschema/ComponentProperties.java
@@ -57,4 +57,5 @@ public class ComponentProperties {
     public void setUiSchema(ObjectNode uiSchema) {
         this.uiSchema = uiSchema;
     }
+
 }
