+++ dataprep-backend-service/src/main/resources/org/talend/dataprep/error_messages.properties
@@ -438,3 +438,5 @@ PREPARATION_NOT_EMPTY.MESSAGE=Preparation {0} is not empty
 FORBIDDEN_PREPARATION_CREATION.TITLE=Preparation error
 FORBIDDEN_PREPARATION_CREATION.MESSAGE=You are not allowed to create a preparation in this folder
 
+BAD_DATASTORE_CONNECTION.TITLE=Connection error
+BAD_DATASTORE_CONNECTION.MESSAGE=Could not connect to data store: {0}
